updating

+++ util/doxygen/process_source_files.py
@@ -11,6 +11,14 @@ import re
 
 def add_doxygen_group (path, group_name):
 
+    if not os.path.exists (path):
+        raise Exception("Cannot add Doxygen group to nonexistent filepath!")
+        return
+
+    if not group_name:
+        raise Exception("Doxygen group name cannot be empty!")
+        return
+
     with open (path, "r") as f:
         content = f.read()
 
@@ -27,6 +35,10 @@ def add_doxygen_group (path, group_name):
 
 def process_module_header (header_path):
 
+    if not os.path.exists (header_path):
+        raise Exception("Cannot parse nonexistent module header!")
+        return
+
     with open (header_path, "r") as f:
         block_info_result = re.match (r".*BEGIN_JUCE_MODULE_DECLARATION"
                                       "(.*)"
@@ -34,13 +46,12 @@ def process_module_header (header_path):
                                       f.read(),
                                       re.DOTALL)
 
-    # The module header causes problems for Doxygen, so delete it
-    os.remove (header_path)
+    os.remove (header_path) # The module header causes problems for Doxygen, so delete it
 
     detail_lines = []
 
-    # read lines from the module description block
     for line in block_info_result.group (1).split ("\n"):
+
         stripped_line = line.strip()
         
         if stripped_line:
@@ -62,6 +73,10 @@ def create_module_subgroup (module_name, module_dir, subdir):
 
     subgroup_name = subdir.split(os.path.sep)[-1]
 
+    if not subgroup_name:
+        raise Excpetion("Error creating module subgroup - empty deduced subgroup name")
+        return
+
     subgroup_definition = []
     subgroup_definition.append ("/** @defgroup {n} {n}".format (n=subgroup_name))
     subgroup_definition.append ("")
@@ -71,10 +86,8 @@ def create_module_subgroup (module_name, module_dir, subdir):
     # put this subdir's contents into appropriate doxygen subgroups
     for dirpath, dirnames, filenames in os.walk (os.path.join (module_dir, subdir)):
         for filename in filenames:
-            # top-level files in this directory
             add_doxygen_group (os.path.join (dirpath, filename), subgroup_name)
         for dirname in dirnames:
-            # nested subdirectories
             subgroup_definition.append ("")
             subgroup_definition.append (create_module_subgroup (module_name, module_dir, os.path.join (subdir, dirname)))
 
@@ -88,10 +101,21 @@ def create_module_subgroup (module_name, module_dir, subdir):
 
 # Processes a JUCE module and returns a module description for Doxygen
 def process_juce_module (category_name, module_name, module_dir):
+
+    if not category_name:
+        raise Exception("Juce module category name cannot be empty!")
+        return
+
+    if not module_name:
+        raise Exception("Juce module name cannot be empty!")
+        return
+
+    if not os.path.isdir (module_dir):
+        raise Exception("Error - juce module directory does not exist!")
+        return
     
     module_header_info = process_module_header (os.path.join (module_dir, module_name + ".h"))
 
-    # Create a Doxygen group definition for the module
     module_definiton = []
     module_definiton.append ("/** @defgroup {n} {n}".format (n=module_name))
     module_definiton.append ("    {d}".format (d=module_header_info[0]))
@@ -100,18 +124,18 @@ def process_juce_module (category_name, module_name, module_dir):
     for line in module_header_info[1]:
         module_definiton.append ("    - {l}".format (l=line))
 
+    del module_header_info
+
     module_definiton.append ("")
     module_definiton.append ("    @{")
     module_definiton.append ("*/")
 
     # put the module's contents into appropriate doxygen subgroups
     for item in os.listdir (module_dir):
-        if (os.path.isdir (os.path.join (module_dir, item))):
-            # subdirectory within module
+        if os.path.isdir (os.path.join (module_dir, item)):
             module_definiton.append ("")
             module_definiton.append (create_module_subgroup (module_name, module_dir, item))
         else:
-            # top-level files within module
             add_doxygen_group (os.path.join (module_dir, item), module_name)
 
     module_definiton.append ("")
@@ -126,6 +150,10 @@ def process_juce_module (category_name, module_name, module_dir):
 
 def process_category_description (file_path):
 
+    if not os.path.exists (file_path):
+        raise Exception("Category desription file does not exist!")
+        return
+
     with open (file_path, "r") as f:
         content = f.read()
 
@@ -137,20 +165,25 @@ def process_category_description (file_path):
 
 def process_module_category (category_name, orig_cat_dir, dest_cat_dir):
 
+    if not os.path.isdir (orig_cat_dir):
+        raise Exception("Module category dir does not exist!")
+        return
+
+    if not category_name:
+        raise Exception("Module category name cannot be empty!")
+        return
+
     # copy files to doxygen working tree
     shutil.copytree (orig_cat_dir, dest_cat_dir)
 
-    category_description = process_category_description (os.path.join (dest_cat_dir, category_name.lower() + ".txt"))
-
-    # Create a Doxygen group definition for the category
     category_definiton = []
     category_definiton.append ("/** @defgroup {n} {n}".format (n=category_name))
-    category_definiton.append ("    {d}".format (d=category_description))
+    category_definiton.append ("    {d}".format (d=process_category_description (os.path.join (dest_cat_dir, 
+                                                                                 category_name.lower() + ".txt"))))
     category_definiton.append ("")
     category_definiton.append ("    @{")
     category_definiton.append ("*/")
 
-    # process all the modules in this category
     for subdir in os.listdir  (dest_cat_dir):
         module_path = os.path.join (dest_cat_dir, subdir)
         if os.path.isdir (module_path):
@@ -169,7 +202,15 @@ def process_module_category (category_name, orig_cat_dir, dest_cat_dir):
 
 def process_juce_modules (source_dir, dest_dir):
 
+    if not os.path.isdir (source_dir):
+        raise Exception("Juce modules source directory does not exist!")
+        return
+
     orig_module_dir = os.path.join (source_dir, "modules")
+
+    if not os.path.isdir (orig_module_dir):
+        raise Exception("Juce modules source directory does not exist!")
+        return
 
     category_definitions = []
 
@@ -193,6 +234,11 @@ def copy_cmake_readme (script_dir, dest_dir):
     filename = "cmake_api.md"
 
     orig_file_path = os.path.join (script_dir, filename)
+
+    if not os.path.exists (orig_file_path):
+        raise Exception("CMake API readme not found!")
+        return
+
     dest_file_path = os.path.join (dest_dir, filename)
 
     shutil.copy2 (orig_file_path, dest_file_path)
@@ -204,9 +250,15 @@ def copy_from_main_readme (source_dir, script_dir, dest_dir):
 
     filename = "main_page.md"
 
+    orig_file_path = os.path.join (script_dir, filename)
+
+    if not os.path.exists (orig_file_path):
+        raise Exception("Main Lemons readme not found!")
+        return
+
     dest_file_path = os.path.join (dest_dir, filename)
 
-    shutil.copy2 (os.path.join (script_dir, filename), dest_file_path)
+    shutil.copy2 (orig_file_path, dest_file_path)
 
     with open (os.path.join (source_dir, "README.md"), "r") as f:
         orig_readme_text = f.read()
@@ -214,6 +266,10 @@ def copy_from_main_readme (source_dir, script_dir, dest_dir):
     last_bit_of_intro = "whatever floats your git boat."
 
     relevant_text = orig_readme_text.split (last_bit_of_intro, 1)[1]
+
+    if not relevant_text:
+        raise Exception("No relevant text found in main Lemons readme")
+        return
 
     with open (dest_file_path, "a") as f:
         f.write (relevant_text)
@@ -230,7 +286,13 @@ def copy_logo (script_dir, dest_dir):
     util_dir = os.path.abspath (os.path.dirname (script_dir))
     assets_dir = os.path.join (util_dir, "assets")
 
-    shutil.copy2 (os.path.join (assets_dir, filename), 
+    orig_logo_file = os.path.join (assets_dir, filename)
+
+    if not os.path.exists (orig_logo_file):
+        raise Exception("Logo not found")
+        return
+
+    shutil.copy2 (orig_logo_file, 
                   os.path.join (dest_dir, filename))
 
 
@@ -242,21 +304,19 @@ if __name__ == "__main__":
 
     script_dir = os.path.abspath (os.path.dirname (__file__))
 
-    source_dir = os.path.abspath (os.path.dirname (os.path.dirname (script_dir)))
-
     dest_dir = os.path.join (script_dir, "build")
 
-    # delete the working dir if it exists
     if os.path.isdir (dest_dir):
         shutil.rmtree (dest_dir)
 
-    # re-create a clean working directory
     os.mkdir (dest_dir)
 
     copy_cmake_readme (script_dir, dest_dir)
 
-    copy_from_main_readme (source_dir, script_dir, dest_dir)
-
     copy_logo (script_dir, dest_dir)
 
-    process_juce_modules (source_dir, dest_dir)
+    lemons_root = os.path.abspath (os.path.dirname (os.path.dirname (script_dir)))
+
+    copy_from_main_readme (lemons_root, script_dir, dest_dir)
+
+    process_juce_modules (lemons_root, dest_dir)
